requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
streamlit==1.29.0
altair==5.2.0
kafka-python-ng==2.2.2
python-dotenv==1.0.0
openpyxl==3.1.2
pyarrow==14.0.1
protobuf==4.25.1